Move tag into new dir
