+++ code/tag_test.go
@@ -1,4 +1,4 @@
-package main
+package code
 
 import (
 	"errors"
